+++ src/common/decorators/index.ts
@@ -1,5 +1,4 @@
-export * from './injectAccessUser.decorator';
+export * from './injectUser.decorator';
 export * from './usePagination.decorator';
-export * from './injectLocalUser.decorator';
-export * from './injectRefreshUser.decorator';
 export * from './namedController';
+export * from './apiKudogExceptionResponse.decorator';
